+++ src/coastline.jl
@@ -148,3 +148,36 @@ function plot_coastline!(coastline::Coastline;
         tickdirection=tickdirection, kwargs...)
 end
 
+"""
+    scale_bar(distance::Real=100.0, x=:topleft, y=:topleft)
+
+Add a scalebar to a plot made with [`plot_coastline`]@ref).
+
+The length of the scalebar, in km, is given by `distance`. The position is
+given by `x` and `y` which, in this preliminary version of the function, must
+both equal `:topleft`.
+
+# Examples
+
+```juliadoc
+using OceanAnalysis, Plots
+cl = coastline();
+plot_coastline(cl, xlim=(-70, -60), ylim=(42, 48), aspect_ratio=1 / cos(45 * pi / 180))
+scale_bar(100.0)
+```
+"""
+function scale_bar(distance::Real=100.0, x=:topleft, y=:topleft)
+    (x == :topleft && y == :topleft) || error("x and y must both be :topleft")
+    xlim = xlims()
+    ylim = ylims()
+    ymid = (ylim[1] + ylim[2]) / 2.0
+    km_per_degree_lon = geod_distance(xlim[1] - 0.5, ymid, xlim[1] + 0.5, ymid)
+    dx = (xlim[2] - xlim[1]) / 20
+    dy = (ylim[2] - ylim[1]) / 20
+    y0 = ylim[2] - dy
+    X = xlim[1] + dx .+ [0.0, distance / km_per_degree_lon]
+    Y = [y0, y0]
+    plot!(X, Y, color=:black, linewidth=2)
+    annotate!((X[1] + X[2]) / 2.0, y0 + dy / 2,
+        Plots.text("$(trunc(Int, distance)) km", 8))
+end