+++ src/setup.jl
@@ -51,6 +51,7 @@ export read_ctd_cnv
 export read_topography
 export salinity_from_conductivity
 export SA
+export scale_bar
 export subset_amsr
 export T90_from_T48
 export T90_from_T68